Refactor createStreamingServer so that chunks are injected by the test.
