+++ src/index.ts
@@ -1,3 +1,3 @@
 export { streamQuery } from "./client";
 export * from "./types";
-export { server as mockServer } from "./mocks/server";
+export { createStreamingServer } from "./createStreamingServer";
